+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Fixed js error in the filterDate example.
+
+4.1.0 - (March 15, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+4.0.0 - (May 13, 2019)
+------------------
 ### Breaking Changes
 * Removed `releaseFocus` prop
 * Removed `requestFocus` prop